requirements.txt
asn1crypto==1.5.1
boto3==1.42.44
botocore==1.42.44
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cloudpickle==3.1.1
cryptography==46.0.4
filelock==3.20.3
idna==3.11
jmespath==1.1.0
packaging==26.0
platformdirs==4.5.1
protobuf==6.33.5
pycparser==3.0
PyJWT==2.11.0
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
s3transfer==0.16.0
setuptools==82.0.0
six==1.17.0
snowflake-connector-python==4.2.0
snowflake-snowpark-python==1.45.0
sortedcontainers==2.4.0
tomlkit==0.14.0
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
wheel==0.46.3
